mcp_server/pyproject.toml
[project]
name = "ai-discovery-mcp"
version = "0.1.0"
description = "MCP server for integrating AI Discovery Documentation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "httpx>=0.28.1",
    "mcp[cli]>=1.3.0",
    "mem0ai>=0.1.88",
    "pygithub>=2.6.1",
    "requests>=2.32.3",
    "vecs>=0.4.5",
]
